Add a link to Google Sky's web edition.
For now, this is only a simple link to the site. It's missing
"longitude" and "latitude" parameters (-> RA/Dec) for now and
the KMZ file cannot be given in the URL either. However, this
is work in progress: I asked internal Google staff for this
to be added. Hope to get a positive answer!